ice: add basic skeleton and TLV framework for live migration support
In preparation for supporting VF live migration, introduce a new functional
block to the ice driver to expose hooks for migrating a VF.

The ice-vfio-pci driver will call ice_migration_init_dev() to indicate that
the migration driver is loaded.

When the host wants to migrate the device, first it suspends the device by
calling ice_migration_suspend_dev(). Then it saves any device state by
calling ice_migration_save_devstate() which serializes the VF device state
into a buffer payload which is sent to the resuming host. This target
system calls ice_migration_suspend_dev() to stop the target VF device, and
finally calls ice_migration_load_devstate() to deserialize the migration
state and program the target VF.

Add the virt/migration.c file and implement the skeleton for these
functions. A basic framework is implemented, but the following patches will
fully flesh out the serialization and deserialization for the various kinds
of VF data.

Previous implementations of live migration were implemented by serializing
device state as a series of virtchnl messages. It was thought that this
would be simple since virtchnl is the ABI for communication between the PF
and VF.

Unfortunately, use of virtchnl presented numerous problems. To avoid these
issues, the migration buffer will be implemented using a custom
Type-Length-Value format which avoids a few key issues with the virtchnl
solution:

1. Migration data is only captured when a live migration event is
   requested. The driver no longer caches a copy of every virtchnl message
   sent by the VF.
2. Migration data size is bounded, and the VF cannot indirectly increase it
   to arbitrary sizes by sending unexpected sequences of virtchnl messages.
3. Replay of VF state is controlled precisely, and no longer requires
   modification of the standard virtchnl communication flow.
4. Additional data about the VF state is sent over the migration buffer,
   which is not captured by virtchnl messages. This includes host state
   such as the VF trust flag, MSI-X configuration, RSS configuration, etc.

Introduce the initial support for this TLV format along with the first TLV
for storing basic VF info.

The TLV definitions are placed in virt/migration_tlv.h which defines the
payload data and describes the ABI for this communication format. The first
TLV must be the ICE_MIG_TLV_HEADER which consists of a magic number and a
version used to identify the payload format. This allows for the
possibility of future extension should the entire format need to be
changed.

TLVs are specified as a series of ice_migration_tlv_structures, which are
variable length structures using a flexible array of bytes. These
structures are __packed. However, access to unaligned memory requires the
compiler to insert additional instructions to avoid unaligned access on
some platforms. To minimize -- but not completely remove -- this, the
length for all TLVs is required to be aligned to 4-bytes. The header itself
is 4 bytes, and this ensures that the header and all values with a size 4
bytes or smaller will have aligned access. 8 byte values may potentially be
unaligned, but use if the __packed attribute ensures that the compiler will
insert appropriate access instructions for all platforms.

Note that this migration implementation generally assumes the host and
target system are of the same endianness, integer size, etc. The chosen
magic number of 0xE8000001 will catch byte order difference, and all types
in the TLVs are fixed size.

The type of the  TLVs is specified by the ice_migration_tlvs enumeration.
This *is* ABI, and any extension must add new TLVs at the end of this list
just prior to the NUM_ICE_MIG_TLV which specifies the number of recognized
TLVs by this version of the ABI.

The exact "version" of the ABI is specified by a combination of the magic
number, the version in the header, and the number of TLVs as specified by
NUM_ICE_MIG_TLV, which is sent in the header. This allows for new TLVs to
be added in the future without needing to increment the version number.
Such an increment should only be done if any existing TLV format must
change, or if the entire format must change for some reason.

The payload *must* begin with the ICE_MIG_TLV_HEADER, as this specifies the
format. All other TLVs do not have an order, and receiving code must be
capable of handling TLVs in an arbitrary order. Some TLVs will be sent
multiple times, such as for Tx and Rx queue information.

The special ICE_MIG_TLV_END type must come last and is a marker to indicate
the end of the payload buffer. It has a size of 0.

Implement a few macros to help reduce boiler plate. The ice_mig_tlv_type()
macro converts a structure pointer to its TLV type using _Generic(). The
ice_mig_alloc_tlv() macro allocates a new TLV, returning the pointer to the
value structure. This macro ensures the 4-byte alignment required for all
TLV lengths. The ice_mig_alloc_flex_tlv() is similar, but allows allocating
a TLV whose element structure ends in a flexible array. Finally, the
ice_mig_tlv_add_tail() macro takes a pointer to an element structure and
inserts the container ice_mig_tlv_entry and inserts it into the TLV list
used for temporary storage.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>