lib: packing: create __pack() and __unpack() variants without error checking
A future variant of the API, which works on arrays of packed_field
structures, will make most of these checks redundant. The idea will be
that we want to perform sanity checks only once at boot time, not once
for every function call. So we need faster variants of pack() and
unpack(), which assume that the input was pre-sanitized.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>