lib: packing: create variants of struct packed_field with reduced sizes
WIP

Using size_t (unsigned long) for the struct packed_field members is too
much in many cases. We can get away even with u8 for small buffers, and
that saves rodata in the packed_field array definitions.

On the other hand, the pack_fields() and unpack_fields() API is
currently strongly typed. We could add pack_fields_8(), pack_fields_16(),
etc, which operate on reduced versions of struct packed_field arrays, or
we can reimplement pack_fields() and unpack_fields() as macros which are
type-agnostic, in a poor man's C++ template sort of way.

The tradeoff with the macro approach is that now, the loop is unrolled
for each pack_fields_m() call, which increases the kernel text size.
How much that matters probably depends on the usage pattern. Few call
sites with a lot of fields will probably benefit from the loop
unrolling, while many call sites with few fields will not. For the
moment, keep both the regular function calls and the macro variants,
until we benchmark and see which one wins out.

The _m naming variant stands for "macro". The size_t variant of struct
packed_field is kept for compatibility with the pack_fields() function.

Co-developed-by: Jacob Keller <jacob.e.keller@intel.com>
Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>